backend/requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
sqlalchemy==1.4.41
sqlmodel==0.0.8
alembic==1.12.1
pydantic==1.10.12
python-dotenv==1.0.0
websockets==12.0

streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
